requirements.txt
flask
openai
langchain
Flask-SocketIO
pypdf==3.16.1
tiktoken
faiss-cpu
youtube-transcript-api==0.6.1
langchain-experimental
neo4j
wikipedia
python-dotenv
